+++ lib/procon_bypass_man.rb
@@ -16,6 +16,7 @@ require_relative "procon_bypass_man/support/on_memory_cache"
 require_relative "procon_bypass_man/support/http_client"
 require_relative "procon_bypass_man/support/report_http_client"
 require_relative "procon_bypass_man/support/update_remote_pbm_action_status_http_client"
+require_relative "procon_bypass_man/support/server_pool"
 require_relative "procon_bypass_man/background"
 require_relative "procon_bypass_man/commands"
 require_relative "procon_bypass_man/bypass"
