+++ lib/procon_bypass_man/configuration.rb
@@ -98,12 +98,12 @@ class ProconBypassMan::Configuration
 
   # @return [Array<ProconBypassMan::ServerPool>]
   def internal_server_pool
-    @server_pool ||= ProconBypassMan::Background::ServerPool.new(servers: internal_api_servers)
+    @server_pool ||= ProconBypassMan::ServerPool.new(servers: internal_api_servers)
   end
 
   # @return [Array<ProconBypassMan::ServerPool>]
   def server_pool
-    @server_pool ||= ProconBypassMan::Background::ServerPool.new(servers: api_servers)
+    @server_pool ||= ProconBypassMan::ServerPool.new(servers: api_servers)
   end
 
   # @return [Array<String>]
